separate baker-compiler-kotlin maven module
