+++ core/baker-compiler-kotlin/src/main/kotlin/com/ing/baker/compiler/ScalaConversions.kt
