+++ core/baker-compiler-kotlin/src/main/kotlin/com/ing/baker/recipe/CheckPointEvent.kt
