+++ core/baker-compiler-kotlin/src/main/kotlin/com/ing/baker/recipe/InteractionDescriptor.kt
